# api-specs/openapi.yaml
openapi: 3.0.3
info:
  title: ESG & Environmental Data Management Platform API
  description: |
    REST API for the ESG Environmental Data Management Platform.
    Manages energy, water, waste, GHG emissions, air emissions, sales, recovery,
    and production data across manufacturing sites.
  version: 2026-02-24
  contact:
    name: ESG Platform Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api/2026-02-24
    description: Local development
  - url: http://localhost:3000/api/2026-02-24
    description: Docker (via nginx proxy)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation error
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    SiteRef:
      type: object
      properties:
        name:
          type: string
        code:
          type: string

    # ── Data Entry Common ──
...
    EnergyData:
      type: object
      required: [siteId, month, year]
      properties:
        id:
          type: integer
          readOnly: true
        siteId:
          type: integer
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
          minimum: 2000
          maximum: 2100
        electricityKwh:
          type: number
          minimum: 0
          default: 0
        steamMt:
          type: number
          minimum: 0
          default: 0
        coalMt:
          type: number
          minimum: 0
          default: 0
        dieselLitres:
          type: number
          minimum: 0
          default: 0
        foLitres:
          type: number
          minimum: 0
          default: 0
        lpgKg:
          type: number
          minimum: 0
          default: 0
        pngScm:
          type: number
          minimum: 0
          default: 0
        renewableKwh:
          type: number
          minimum: 0
          default: 0
        totalEnergyGj:
          type: number
          readOnly: true
        site:
          $ref: '#/components/schemas/SiteRef'
...
    AirEmissionsData:
      type: object
      required: [siteId, month, year]
      properties:
        id:
          type: integer
          readOnly: true
        siteId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        soxKg:
          type: number
          minimum: 0
        noxKg:
          type: number
          minimum: 0
        pmKg:
          type: number
          minimum: 0
        vocKg:
          type: number
          minimum: 0
        stackId:
          type: string
          maxLength: 50
        site:
          $ref: '#/components/schemas/SiteRef'
...
    DashboardSummary:
      type: object
      properties:
        year:
          type: integer
        siteId:
          oneOf:
            - type: integer
            - type: string
              enum: ['all']
        totalScope1:
          type: number
        totalScope2:
          type: number
        totalScope3:
          type: number
        totalGroupEmissions:
          type: number
        carbonIntensity:
          type: number
        energyIntensity:
          type: number
        waterIntensity:
          type: number
        wasteIntensity:
          type: number
        renewableEnergyPct:
          type: number
        recoveryPct:
          type: number
        productionTotal:
          type: number
        totalEnergyGj:
          type: number
        totalWaterKl:
          type: number
        totalWasteMt:
          type: number
    SiteComparison:
      type: object
      properties:
        siteId:
          type: integer
        siteName:
          type: string
        siteCode:
          type: string
        totalEnergyGj:
          type: number
        productionMt:
          type: number
        totalEmissions:
          type: number
    EmissionsTrend:
      type: object
      properties:
        month:
          type: integer
        totalEnergyGj:
          type: number
        electricityKwh:
          type: number
        renewableKwh:
          type: number
...
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Duplicate record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationFailed:
      description: Request body validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
security:
  - bearerAuth: []
paths:
...
  /data/energy:
    get:
      tags: [Energy]
      summary: List energy data
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of energy records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnergyData'
    post:
      tags: [Energy]
      summary: Create energy entry (auto-calculates total GJ)
      description: Only admin and site_user roles can create entries.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnergyData'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnergyData'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /data/energy/{id}:
    put:
      tags: [Energy]
      summary: Update energy entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnergyData'
      responses:
        '200':
          description: Updated record
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Energy]
      summary: Delete energy entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Production ──
...
  /data/air-emissions:
    get:
      tags: [Air Emissions]
      summary: List air emissions data
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of air emissions records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AirEmissionsData'
    post:
      tags: [Air Emissions]
      summary: Create air emissions entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AirEmissionsData'
      responses:
        '201':
          description: Created
  /data/air-emissions/{id}:
    put:
      tags: [Air Emissions]
      summary: Update air emissions entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Air Emissions]
      summary: Delete air emissions entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Sales ──
...
  /dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Get KPI summary metrics
      parameters:
        - name: year
          in: query
          schema:
            type: integer
          description: Defaults to current year
        - name: siteId
          in: query
          schema:
            type: integer
          description: Filter by site (omit for all sites)
      responses:
        '200':
          description: Summary KPIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
  /dashboard/site-comparison:
    get:
      tags: [Dashboard]
      summary: Site-by-site comparison
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of site comparisons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SiteComparison'
  /dashboard/emissions-trend:
    get:
      tags: [Dashboard]
      summary: Monthly energy/emissions trend
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: siteId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 12-month trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmissionsTrend'

  # ── Users ──
...